pull out tests/common
