+++ storage/durability/tests/basic.rs
@@ -18,35 +18,12 @@
 #![deny(unused_must_use)]
 #![deny(rust_2018_idioms)]
 
-use durability::{wal::WAL, DurabilityRecord, DurabilityRecordType, DurabilityService, RawRecord};
+use durability::{DurabilityRecord, DurabilityService, RawRecord};
 use tempdir::TempDir;
 
-#[derive(Debug, PartialEq, Eq, Clone)]
-struct TestRecord {
-    bytes: Vec<u8>,
-}
+mod common;
 
-impl DurabilityRecord for TestRecord {
-    const RECORD_TYPE: DurabilityRecordType = 0;
-    const RECORD_NAME: &'static str = "TEST";
-
-    fn serialise_into(&self, writer: &mut impl std::io::Write) -> bincode::Result<()> {
-        writer.write_all(&self.bytes)?;
-        Ok(())
-    }
-
-    fn deserialize_from(reader: &mut impl std::io::Read) -> bincode::Result<Self> {
-        let mut bytes = Vec::new();
-        reader.read_to_end(&mut bytes)?;
-        Ok(Self { bytes })
-    }
-}
-
-fn open_wal(directory: &TempDir) -> WAL {
-    let mut wal = WAL::open(directory).unwrap();
-    wal.register_record_type::<TestRecord>();
-    wal
-}
+use self::common::{open_wal, TestRecord};
 
 #[test]
 fn basic() {
